split ResultOwner file (no code changes)
